README.md
Tasked with creating a simple information system for an HVAC Company in my Database Design Course at University OF Baltimore. It displays the contents of each entity and the relationships as well. It Is Complete with CREATE & INSERT statements, As Well as Comments for each entity and the information that belongs in them. NOTE: INFORMATION INSERTED IN THIS DATABASE IS NOT REAL!!!